DynamicTablesPkg: AcpiSsdtPcieLibArm: Create support library
Add support library to allow for customization of _OSC and slot info.
The functions in the library are unchanged,
with the exception of adding PciInfo pointer to the APIs.

Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/edk2/+/2742069
GVS: Gerrit_Virtual_Submit
