+++ DynamicTablesPkg/Library/Common/SsdtPcieSupportLib/SsdtPcieSupportLib.inf
@@ -1,5 +1,5 @@
 ## @file
-# Ssdt Serial Port Table Generator
+# Ssdt Serial Port Table Support Library
 #
 #  Copyright (c) 2021, Arm Limited. All rights reserved.<BR>
 #
@@ -8,17 +8,16 @@
 
 [Defines]
   INF_VERSION    = 0x0001001B
-  BASE_NAME      = SsdtPcieLibArm
-  FILE_GUID      = E431D7FD-26BF-4E3D-9064-5B13B0439057
+  BASE_NAME      = SsdtPcieSupportLib
+  FILE_GUID      = 510451a0-60b2-446c-b6bf-59cbe4a41782
   VERSION_STRING = 1.0
   MODULE_TYPE    = DXE_DRIVER
-  LIBRARY_CLASS  = NULL|DXE_DRIVER
-  CONSTRUCTOR    = AcpiSsdtPcieLibConstructor
-  DESTRUCTOR     = AcpiSsdtPcieLibDestructor
+  LIBRARY_CLASS  = SsdtPcieSupportLib
 
 [Sources]
-  SsdtPcieGenerator.c
-  SsdtPcieGenerator.h
+  SsdtPcieSupportLib.c
+  SsdtPcieSupportLibPrivate.h
+  SsdtPcieOscTemplate.asl
 
 [Packages]
   DynamicTablesPkg/DynamicTablesPkg.dec
@@ -29,4 +28,3 @@
   AcpiHelperLib
   AmlLib
   BaseLib
-  SsdtPcieSupportLib
